fix(runtime): use the native KVM ioctl request type

+++ elastos/crates/elastos-crosvm/src/lib.rs
@@ -65,7 +65,7 @@ fn kvm_api_available(path: &std::path::Path) -> bool {
     // Linux UAPI: _IO(KVMIO, 0x00), where KVMIO is 0xAE. This read-only ioctl
     // takes no third argument and returns the API version; it creates no VM.
     // Version 12 is the stable API required by KVM userspace implementations.
-    unsafe { libc::ioctl(device.as_raw_fd(), 0xAE00 as libc::c_ulong) == 12 }
+    unsafe { libc::ioctl(device.as_raw_fd(), 0xAE00) == 12 }
 }
 
 #[cfg(test)]
